Refactor metadata facing routes into separate trait
