+++ engine/src/main/scala/cromwell/webservice/SwaggerService.scala
@@ -1,5 +1,7 @@
 package cromwell.webservice
 
+import cromwell.webservice.routes.CromwellApiService
+
 trait SwaggerService extends SwaggerUiResourceHttpService {
   override def swaggerServiceName = "cromwell"
