+++ engine/src/main/scala/cromwell/webservice/WorkflowJsonSupport.scala
@@ -11,7 +11,7 @@ import cromwell.util.JsonFormatting.WomValueJsonFormatter
 import WomValueJsonFormatter._
 import better.files.File
 import cromwell.services.healthmonitor.HealthMonitorServiceActor.{StatusCheckResponse, SubsystemStatus}
-import cromwell.webservice.CromwellApiService.BackendResponse
+import cromwell.webservice.routes.CromwellApiService.BackendResponse
 import cromwell.webservice.metadata.MetadataBuilderActor.BuiltMetadataResponse
 import spray.json.{DefaultJsonProtocol, JsString, JsValue, RootJsonFormat}
