+++ engine/src/test/scala/cromwell/webservice/SwaggerUiHttpServiceSpec.scala
@@ -6,6 +6,7 @@ import akka.http.scaladsl.server.Route
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import com.typesafe.config.ConfigFactory
 import cromwell.webservice.SwaggerUiHttpServiceSpec._
+import cromwell.webservice.routes.CromwellApiService
 import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
